requirements.txt
langchain
langchain-community
langchain-groq
langchain-text-splitters
langchain-chroma
langchain-huggingface
chromadb
transformers
sentence-transformers
streamlit
python-dotenv
unstructured
